requirements.txt
Flask>=2.0
psycopg2-binary
python-dotenv